Refactor DICOMPlugin.
